refactor: name interval hoo file name
